# swagger.yml
openapi: 3.0.0
paths:
...
  /api/locations:
    get:
      operationId: LocationProxyController_list
      parameters:
        - name: search
          required: false
          in: query
          description: Search by location name or code
          schema:
            example: Pharmacy
            type: string
        - name: limit
          required: false
          in: query
          description: Items per page
          schema:
            example: 20
            type: number
        - name: page
          required: false
          in: query
          description: Page number (1-based)
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: ''
      summary: Search locations proxied from the identity service
      tags: &ref_1
        - location-proxy
  /api/locations/{id}:
    get:
      operationId: LocationProxyController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get a location by id from the identity service
      tags: *ref_1
  /api/patients:
    get:
      operationId: PatientsController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List patients with pagination and search
      tags: &ref_2
        - patients
    post:
      operationId: PatientsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientDto'
      responses:
        '201':
          description: ''
      summary: Register a new patient
      tags: *ref_2
  /api/patients/by-mrn/{patientId}:
    get:
      operationId: PatientsController_getByMrn
      parameters:
        - name: patientId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get patient by MRN (patientId)
      tags: *ref_2
  /api/patients/{id}:
    get:
      operationId: PatientsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get patient by id
      tags: *ref_2
    patch:
      operationId: PatientsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientDto'
      responses:
        '200':
          description: ''
      summary: Update a patient
      tags: *ref_2
    delete:
      operationId: PatientsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a patient
      tags: *ref_2
  /api/payment-providers:
    get:
      operationId: PaymentProvidersController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            allOf:
              - $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            default: 20
            allOf:
              - $ref: '#/components/schemas/Object'
        - name: search
          required: false
          in: query
          description: Search query
          schema:
            type: string
        - name: sortBy
          required: false
          in: query
          schema:
            default: createdAt
            allOf:
              - $ref: '#/components/schemas/Object'
        - name: sortOrder
          required: false
          in: query
          schema:
            default: desc
            type: string
            enum:
              - asc
              - desc
        - name: filter
          required: false
          in: query
          description: Generic filter string
          schema:
            type: string
        - name: organizationId
          required: false
          in: query
          description: Filter by organization (superadmin only)
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: List payment providers with pagination and search
      tags: &ref_3
        - payment-providers
    post:
      operationId: PaymentProvidersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentProviderDto'
      responses:
        '201':
          description: ''
      summary: Create a payment provider
      tags: *ref_3
  /api/payment-providers/{id}:
    get:
      operationId: PaymentProvidersController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get payment provider by id
      tags: *ref_3
    patch:
      operationId: PaymentProvidersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePaymentProviderDto'
      responses:
        '200':
          description: ''
      summary: Update a payment provider
      tags: *ref_3
    delete:
      operationId: PaymentProvidersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a payment provider
      tags: *ref_3
...
  /api/form-submissions:
    get:
      operationId: FormSubmissionsController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List form submissions (filter by patient/visit/encounter/form)
      tags: &ref_14
        - form-submissions
    post:
      operationId: FormSubmissionsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFormSubmissionDto'
      responses:
        '201':
          description: ''
      summary: Submit (or draft) a form against a patient/encounter
      tags: *ref_14
  /api/form-submissions/{id}:
    get:
      operationId: FormSubmissionsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get form submission by id
      tags: *ref_14
    patch:
      operationId: FormSubmissionsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFormSubmissionDto'
      responses:
        '200':
          description: ''
      summary: Update a draft submission
      tags: *ref_14
    delete:
      operationId: FormSubmissionsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a form submission
      tags: *ref_14
  /api/form-submissions/{id}/chain:
    get:
      operationId: FormSubmissionsController_getChain
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get the amend chain (original + all amendments) for a submission
      tags: *ref_14
  /api/form-submissions/{id}/pdf:
    get:
      operationId: FormSubmissionsController_pdf
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Render a submission as a PDF document
      tags: *ref_14
  /api/form-submissions/{id}/amend:
    post:
      operationId: FormSubmissionsController_amend
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFormSubmissionDto'
      responses:
        '201':
          description: ''
      summary: Amend a submitted form (creates a new submission)
      tags: *ref_14
...
info:
  title: EMR API
  description: >-
    Electronic Medical Record: appointments, visits, encounters, dynamic forms,
    and clinical requests
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreatePatientDto:
      type: object
      properties:
        patientId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        otherNames:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        nextOfKinName:
          type: string
        nextOfKinPhone:
          type: string
        nextOfKinRelationship:
          type: string
        identifiers:
          type: array
          items:
            type: string
        maritalStatus:
          type: string
        occupation:
          type: string
        bloodGroup:
          type: string
        genotype:
          type: string
        paymentProviderIds:
          type: array
          items:
            type: string
      required:
        - firstName
        - lastName
    UpdatePatientDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        otherNames:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        nextOfKinName:
          type: string
        nextOfKinPhone:
          type: string
        nextOfKinRelationship:
          type: string
        identifiers:
          type: array
          items:
            type: string
        maritalStatus:
          type: string
        occupation:
          type: string
        bloodGroup:
          type: string
        genotype:
          type: string
        paymentProviderIds:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    Object:
      type: object
      properties: {}
    CreatePaymentProviderDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
          enum: &ref_15
            - CASH
            - HMO
            - COMPANY
            - PROGRAM
            - OTHER
        description:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
      required:
        - code
        - name
        - type
    UpdatePaymentProviderDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
          enum: *ref_15
        description:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
        isActive:
          type: boolean
...
    CreateFormSubmissionDto:
      type: object
      properties:
        formDefinitionId:
          type: string
        patientId:
          type: string
        visitId:
          type: string
        encounterId:
          type: string
        dataJson:
          type: object
        status:
          type: string
          enum:
            - DRAFT
            - SUBMITTED
      required:
        - formDefinitionId
        - patientId
        - dataJson
    UpdateFormSubmissionDto:
      type: object
      properties:
        dataJson:
          type: object
        visitId:
          type: string
        encounterId:
          type: string
        status:
          type: string
          enum:
            - DRAFT
            - SUBMITTED
            - AMENDED